Button styles, Button outline, Button size, Block level buttons, Active/Disabled buttons
The button classes can be used on an <a>, <button>, or <input> element.

The class .active makes a button appear pressed, and the disabled attribute makes a button unclickable. Note that <a> elements do not support the disabled attribute and must therefore use the .disabled class to make it visually appear disabled.