Badge within a headings and buttons, Contextual badges, Pill badges.
Badges are used to add additional information to any content. Use the .badge class together with a contextual class (like .badge-secondary) within <span> elements to create rectangular badges. Note that badges scale to match the size of the parent element (if any).

Use any of the contextual classes (.badge-*) to change the color of a badge.

Use the .badge-pill class to make the badges more round.