Basic and colored labelled progress bars, Progress bar width and height
To create a default progress bar, add a .progress class to a container element and add the .progress-bar class to its child element. Use the CSS width property to set the width of the progress bar.

The height of the progress bar is 16px by default. Use the CSS height property to change it. Note that you must set the same height for the progress container and the progress bar.

Add text inside the progress bar to show the visible percentage.

By default, the progress bar is blue (primary). Use any of the Bootstrap 4 contextual background classes to its color.